Add Lambda Layers and data archives
Terraform now generates archives (.zip files) for each Lambda fucntion
from individual files in the code/ directory as well as the base64 of
the SHA-256 hash to confirm correctness of Lambda functions in AWS.
This streamlines deployment and ensures no discrepencies with the
deployed version and the version in source control.

Also moves the anejocommon.py file into its own Lambda Layer, ensuring
it is consistent across all Anejo functions (and only need be deployed
once).